chore: refactor all DLP samples and tests
