+++ samples/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@google-cloud/dlp": "^3.0.2",
     "@google-cloud/pubsub": "^2.0.0",
+    "google-auth-library": "^6.0.6",
     "mime": "^2.3.1",
     "yargs": "^15.0.0"
   },
