+++ samples/system-test/deid.test.js
@@ -16,13 +16,13 @@
 
 const path = require('path');
 const {assert} = require('chai');
-const {describe, it} = require('mocha');
+const {describe, it, before} = require('mocha');
 const fs = require('fs');
 const cp = require('child_process');
+const DLP = require('@google-cloud/dlp');
 
 const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
 
-const cmd = 'node deid.js';
 const harmfulString = 'My SSN is 372819127';
 const harmlessString = 'My favorite color is blue';
 const surrogateType = 'SSN_TOKEN';
@@ -31,27 +31,39 @@ const tempOutputFile = path.join(__dirname, 'temp.result.csv');
 const dateShiftAmount = 30;
 const dateFields = 'birth_date register_date';
 
+const client = new DLP.DlpServiceClient();
 describe('deid', () => {
+  let projectId;
+
+  before(async () => {
+    projectId = await client.getProjectId();
+  });
   // deidentify_masking
   it('should mask sensitive data in a string', () => {
-    const output = execSync(`${cmd} deidMask "${harmfulString}" -m x -n 5`);
+    const output = execSync(
+      `node deidentifyWithMask.js ${projectId} "${harmfulString}" -m x -n 5`
+    );
     assert.include(output, 'My SSN is xxxxx9127');
   });
 
   it('should ignore insensitive data when masking a string', () => {
-    const output = execSync(`${cmd} deidMask "${harmlessString}"`);
+    const output = execSync(
+      `node deidentifyWithMask.js ${projectId} "${harmlessString}"`
+    );
     assert.include(output, harmlessString);
   });
 
   it('should handle masking errors', () => {
-    const output = execSync(`${cmd} deidMask "${harmfulString}" -n -1`);
+    const output = execSync(
+      `node deidentifyWithMask.js ${projectId} "${harmfulString}" -n -1`
+    );
     assert.include(output, 'Error in deidentifyWithMask');
   });
 
   // deidentify_fpe
   it('should handle FPE encryption errors', () => {
     const output = execSync(
-      `${cmd} deidFpe "${harmfulString}" BAD_KEY_NAME BAD_KEY_NAME`
+      `node deidentifyWithDateShift.js ${projectId} "${harmfulString}" BAD_KEY_NAME BAD_KEY_NAME`
     );
     assert.match(output, /Error in deidentifyWithFpe/);
   });
@@ -59,7 +71,7 @@ describe('deid', () => {
   // reidentify_fpe
   it('should handle FPE decryption errors', () => {
     const output = execSync(
-      `${cmd} reidFpe "${harmfulString}" ${surrogateType} BAD_KEY_NAME BAD_KEY_NAME -a NUMERIC`
+      `node reidentifyWithFpe.js ${projectId} "${harmfulString}" ${surrogateType} BAD_KEY_NAME BAD_KEY_NAME -a NUMERIC`
     );
     assert.match(output, /Error in reidentifyWithFpe/);
   });
@@ -68,7 +80,7 @@ describe('deid', () => {
   it('should date-shift a CSV file', () => {
     const outputCsvFile = 'dates.actual.csv';
     const output = execSync(
-      `${cmd} deidDateShift "${csvFile}" "${outputCsvFile}" ${dateShiftAmount} ${dateShiftAmount} ${dateFields}`
+      `node deidentifyWithDateShift.js ${projectId} "${csvFile}" "${outputCsvFile}" ${dateShiftAmount} ${dateShiftAmount} ${dateFields}`
     );
     assert.include(
       output,
@@ -82,7 +94,7 @@ describe('deid', () => {
 
   it('should handle date-shift errors', () => {
     const output = execSync(
-      `${cmd} deidDateShift "${csvFile}" "${tempOutputFile}" ${dateShiftAmount} ${dateShiftAmount}`
+      `node deidentifyWithDateShift.js ${projectId} "${csvFile}" "${tempOutputFile}" ${dateShiftAmount} ${dateShiftAmount}`
     );
     assert.match(output, /Error in deidentifyWithDateShift/);
   });
