+++ samples/system-test/risk.test.js
@@ -19,6 +19,7 @@ const {describe, it, before, after} = require('mocha');
 const uuid = require('uuid');
 const {PubSub} = require('@google-cloud/pubsub');
 const cp = require('child_process');
+const DLP = require('@google-cloud/dlp');
 
 const execSync = cmd => {
   return cp.execSync(cmd, {
@@ -27,12 +28,11 @@ const execSync = cmd => {
   });
 };
 
-const cmd = 'node risk.js';
 const dataset = 'integration_tests_dlp';
 const uniqueField = 'Name';
 const numericField = 'Age';
-const testProjectId = process.env.GCLOUD_PROJECT;
 const pubsub = new PubSub();
+const client = new DLP.DlpServiceClient();
 
 /*
  * The tests in this file rely on a table in BigQuery entitled
@@ -44,11 +44,14 @@ const pubsub = new PubSub();
  * Insert into this table a few rows of Age/Name pairs.
  */
 describe('risk', () => {
+  let projectId;
   // Create new custom topic/subscription
   let topic, subscription;
   const topicName = `dlp-risk-topic-${uuid.v4()}-${Date.now()}`;
   const subscriptionName = `dlp-risk-subscription-${uuid.v4()}-${Date.now()}`;
+
   before(async () => {
+    projectId = await client.getProjectId();
     [topic] = await pubsub.createTopic(topicName);
     [subscription] = await topic.createSubscription(subscriptionName);
     await deleteOldTopics();
@@ -84,7 +87,7 @@ describe('risk', () => {
   // numericalRiskAnalysis
   it('should perform numerical risk analysis', () => {
     const output = execSync(
-      `${cmd} numerical ${dataset} harmful ${numericField} ${topicName} ${subscriptionName} -p ${testProjectId}`
+      `node numericalRiskAnalysis.js ${projectId} ${dataset} harmful ${numericField} ${topicName} ${subscriptionName} -p ${projectId}`
     );
     assert.match(output, /Value at 0% quantile:/);
     assert.match(output, /Value at \d+% quantile:/);
@@ -92,7 +95,7 @@ describe('risk', () => {
 
   it('should handle numerical risk analysis errors', () => {
     const output = execSync(
-      `${cmd} numerical ${dataset} nonexistent ${numericField} ${topicName} ${subscriptionName} -p ${testProjectId}`
+      `node numericalRiskAnalysis.js ${projectId} ${dataset} nonexistent ${numericField} ${topicName} ${subscriptionName} -p ${projectId}`
     );
     assert.match(output, /Error in numericalRiskAnalysis/);
   });
@@ -100,21 +103,21 @@ describe('risk', () => {
   // categoricalRiskAnalysis
   it('should perform categorical risk analysis on a string field', () => {
     const output = execSync(
-      `${cmd} categorical ${dataset} harmful ${uniqueField} ${topicName} ${subscriptionName} -p ${testProjectId}`
+      `node categoricalRiskAnalysis.js ${projectId} ${dataset} harmful ${uniqueField} ${topicName} ${subscriptionName} -p ${projectId}`
     );
     assert.match(output, /Most common value occurs \d time\(s\)/);
   });
 
   it('should perform categorical risk analysis on a number field', () => {
     const output = execSync(
-      `${cmd} categorical ${dataset} harmful ${numericField} ${topicName} ${subscriptionName} -p ${testProjectId}`
+      `node categoricalRiskAnalysis.js ${projectId} ${dataset} harmful ${numericField} ${topicName} ${subscriptionName} -p ${projectId}`
     );
     assert.match(output, /Most common value occurs \d time\(s\)/);
   });
 
   it('should handle categorical risk analysis errors', () => {
     const output = execSync(
-      `${cmd} categorical ${dataset} nonexistent ${uniqueField} ${topicName} ${subscriptionName} -p ${testProjectId}`
+      `node categoricalRiskAnalysis.js ${projectId} ${dataset} nonexistent ${uniqueField} ${topicName} ${subscriptionName} -p ${projectId}`
     );
     assert.match(output, /Error in categoricalRiskAnalysis/);
   });
@@ -122,7 +125,7 @@ describe('risk', () => {
   // kAnonymityAnalysis
   it('should perform k-anonymity analysis on a single field', () => {
     const output = execSync(
-      `${cmd} kAnonymity ${dataset} harmful ${topicName} ${subscriptionName} ${numericField} -p ${testProjectId}`
+      `node kAnonymityAnalysis.js ${projectId} ${dataset} harmful ${topicName} ${subscriptionName} ${numericField} -p ${projectId}`
     );
     assert.match(output, /Quasi-ID values:/);
     assert.match(output, /Class size: \d/);
@@ -130,7 +133,7 @@ describe('risk', () => {
 
   it('should handle k-anonymity analysis errors', () => {
     const output = execSync(
-      `${cmd} kAnonymity ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -p ${testProjectId}`
+      `node kAnonymityAnalysis.js ${projectId} ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -p ${projectId}`
     );
     assert.match(output, /Error in kAnonymityAnalysis/);
   });
@@ -138,7 +141,7 @@ describe('risk', () => {
   // kMapAnalysis
   it('should perform k-map analysis on a single field', () => {
     const output = execSync(
-      `${cmd} kMap ${dataset} harmful ${topicName} ${subscriptionName} ${numericField} -t AGE -p ${testProjectId}`
+      `node kMapEstimationAnalysis.js ${projectId} ${dataset} harmful ${topicName} ${subscriptionName} ${numericField} -t AGE -p ${projectId}`
     );
     assert.match(output, /Anonymity range: \[\d+, \d+\]/);
     assert.match(output, /Size: \d/);
@@ -147,7 +150,7 @@ describe('risk', () => {
 
   it('should handle k-map analysis errors', () => {
     const output = execSync(
-      `${cmd} kMap ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -t AGE -p ${testProjectId}`
+      `node kMapEstimationAnalysis.js ${projectId} ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -t AGE -p ${projectId}`
     );
     assert.match(output, /Error in kMapEstimationAnalysis/);
   });
@@ -155,7 +158,7 @@ describe('risk', () => {
   it('should check that numbers of quasi-ids and info types are equal', () => {
     assert.throws(() => {
       execSync(
-        `${cmd} kMap ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -t AGE GENDER -p ${testProjectId}`
+        `node kMapEstimationAnalysis.js ${projectId} ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -t AGE GENDER -p ${projectId}`
       );
     }, /Number of infoTypes and number of quasi-identifiers must be equal!/);
   });
@@ -163,7 +166,7 @@ describe('risk', () => {
   // lDiversityAnalysis
   it('should perform l-diversity analysis on a single field', () => {
     const output = execSync(
-      `${cmd} lDiversity ${dataset} harmful ${uniqueField} ${topicName} ${subscriptionName} ${numericField} -p ${testProjectId}`
+      `node lDiversityAnalysis.js ${projectId} ${dataset} harmful ${uniqueField} ${topicName} ${subscriptionName} ${numericField} -p ${projectId}`
     );
     assert.match(output, /Quasi-ID values:/);
     assert.match(output, /Class size: \d/);
@@ -172,7 +175,7 @@ describe('risk', () => {
 
   it('should handle l-diversity analysis errors', () => {
     const output = execSync(
-      `${cmd} lDiversity ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -p ${testProjectId}`
+      `node lDiversityAnalysis.js ${projectId} ${dataset} nonexistent ${topicName} ${subscriptionName} ${numericField} -p ${projectId}`
     );
     assert.match(output, /Error in lDiversityAnalysis/);
   });
